Added saving to the result, information about file sizes (bytes megabytes). Displaying size information in the database. Correction of paths depending on the operating system.

+++ audiohandler/database/audioDB.py
@@ -63,9 +63,9 @@ class AudioDB():
                             format TEXT,
                             date DateTime,
                             original_size_b INT,
-                            original_size_mb INT,
+                            original_size_mb FLOAT,
                             convert_size_b INT,
-                            convert_size_mb INT
+                            convert_size_mb FLOAT
                             )"""
 
         with open_db(db) as conn:
